Rename variations declaration model to package models.playback.Declaration
