+++ app/pages/DeclarationChangesNoChangesPage.scala
@@ -16,10 +16,10 @@
 
 package pages
 
-import models.DeclarationChangesNoChanges
+import models.playback.Declaration
 import play.api.libs.json.JsPath
 
-case object DeclarationChangesNoChangesPage extends QuestionPage[DeclarationChangesNoChanges] {
+case object DeclarationChangesNoChangesPage extends QuestionPage[Declaration] {
 
   override def path: JsPath = JsPath \ toString
